requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cbor==1.0.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cmake==3.31.6
datasets==4.0.0
dill==0.3.8
dnspython==2.7.0
filelock==3.18.0
FlagEmbedding==1.3.5
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langsmith==0.4.6
lxml==6.0.0
lz4==4.4.4
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numpy==2.3.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
peft==0.16.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
pyspellchecker==0.8.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
trec-car-tools==2.6
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
zstandard==0.23.0
